api-service/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
pika
python-consul
psycopg2-binary
gunicorn
whitenoise